Bug 600141 - Add -pthread to gmodule pkg-config
It is not generally permissible to dlopen() libpthread into a program
that is not already linked with threading support.  The most common case
of this happening is when GIO opens Gvfs (which includes DBus, which
links against libpthread).

Since gmodule is the way that most users cause this problem for
themselves and since the easiest workaround is to link the program with
-pthread, the best fix for the issue is to add -pthread to the gmodule
pkg-config.